# internal/infrastructure/database/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package dbgen

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Cat struct {
	ID         pgtype.UUID
	Name       string
	Experience int32
	Breed      string
	Salary     int32
}

type Mission struct {
	ID         pgtype.UUID
	Targets    []byte
	AssigneeID pgtype.UUID
	Complete   bool
}